requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
Click==7.0
cycler==0.10.0
decorator==4.3.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
kiwisolver==1.0.1
matplotlib==3.0.2
more-itertools==4.3.0
numpy==1.15.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow
pluggy==0.8.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
py==1.7.0
Pygments==2.3.1
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
pytest==4.0.2
python-dateutil==2.7.5
six==1.12.0
svgwrite==1.2.1
traitlets==4.3.2
wcwidth==0.1.7
